Crate usvg_tree

usvg-tree is an SVG tree representation used by usvg.

Re-exports

pub use strict_num;

pub use tiny_skia_path;

Modules

filter

SVG filter types.

utils

Some useful utilities.

Structs

AspectRatio

Representation of the preserveAspectRatio attribute.

BBox

A bounding box calculator.

BaseGradient

A generic gradient.

ClipPath

A clip-path element.

Color

A 8-bit RGB color.

Fill

A fill style.

Font

Text font properties.

Group

A group container.

Image

A raster image element.

LinearGradient

A linear gradient.

Mask

A mask element.

NonZeroF32

A non-zero f32.

NonZeroPositiveF32

An immutable, finite f32 that is known to be > 0.

NonZeroRect

A rectangle defined by left, top, right and bottom edges.

NormalizedF32

An immutable, finite f32 in a 0..=1 range.

Path

A path element.

Pattern

A pattern element.

PositiveF32

An immutable, finite f32 that is known to be >= 0.

RadialGradient

A radial gradient.

Rect

A rectangle defined by left, top, right and bottom edges.

Size

A size.

Stop

Gradient’s stop element.

Stroke

A stroke style.

StrokeMiterlimit

A stroke-miterlimit value.

Text

A text element.

TextChunk

A text chunk.

TextDecoration

A text span decoration.

TextDecorationStyle

A text span decoration style.

TextPath

A path used by text-on-path.

TextSpan

A text style span.

Transform

An affine transformation matrix.

Tree

A nodes tree container.

ViewBox

View box.

Enums

Align

Representation of the align value of the preserveAspectRatio attribute.

AlignmentBaseline

An alignment baseline property.

BaselineShift

A baseline shift property.

BlendMode

A blending mode property.

DominantBaseline

A dominant baseline property.

FillRule

A fill rule.

FontStretch

A font stretch property.

FontStyle

A font style property.

ImageKind

An embedded image kind.

ImageRendering

An image rendering method.

LengthAdjust

A length adjust property.

LineCap

A line cap.

LineJoin

A line join.

MaskType

A mask type.

NodeKind

Node’s kind.

Paint

A paint style.

PaintOrder

Representation of the paint-order property.

ShapeRendering

A shape rendering method.

SpreadMethod

A spread method.

TextAnchor

A text chunk anchor property.

TextFlow

A text chunk flow property.

TextRendering

A text rendering method.

Units

An element units.

Visibility

A visibility property.

WritingMode

A writing mode.

Traits

ApproxEqUlps

ApproxEqUlps is a trait for approximate equality comparisons. The associated type Flt is a floating point type which implements Ulps, and is required so that this trait can be implemented for compound types (e.g. vectors), not just for the floats themselves.

ApproxZeroUlps

Approximate zero equality comparisons.

IsValidLength

Checks that the current number is > 0.

NodeExt

Additional Node methods.

Type Aliases

Node

Alias for rctree::Node<NodeKind>.

Opacity

An alias to NormalizedF32.

StopOffset

An alias to NormalizedF32.

StrokeWidth

An alias to NonZeroPositiveF32.