Struct usvg_tree::Tree

pub struct Tree {
    pub size: Size,
    pub view_box: ViewBox,
    pub root: Node,
}

A nodes tree container.

Fields
size: Size

Image size.

Size of an image that should be created to fit the SVG.

width and height in SVG.

view_box: ViewBox

SVG viewbox.

Specifies which part of the SVG image should be rendered.

viewBox and preserveAspectRatio in SVG.

root: Node

The root element of the SVG tree.

The root node is always Group.

Implementations

impl Tree

pub fn node_by_id(&self, id: &str) -> Option<Node>

Returns renderable node by ID.

If an empty ID is provided, than this method will always return None. Even if tree has nodes with empty ID.

pub fn has_text_nodes(&self) -> bool

Checks if the current tree has any text nodes.

pub fn paint_servers<F: FnMut(&Paint)>(&self, f: F)

Calls a closure for each Paint in the tree.

Doesnโt guarantee to have unique paint servers. A caller must deduplicate them manually.

pub fn clip_paths<F: FnMut(Rc<ClipPath>)>(&self, f: F)

Calls a closure for each ClipPath in the tree.

Doesnโt guarantee to have unique clip paths. A caller must deduplicate them manually.

pub fn masks<F: FnMut(Rc<Mask>)>(&self, f: F)

Calls a closure for each Mask in the tree.

Doesnโt guarantee to have unique masks. A caller must deduplicate them manually.

pub fn filters<F: FnMut(Rc<Filter>)>(&self, f: F)

Calls a closure for each Filter in the tree.

Doesnโt guarantee to have unique filters. A caller must deduplicate them manually.

pub fn calculate_abs_transforms(&mut self)

Calculates absolute transforms for all nodes in the tree.

As of now, sets Group::abs_transform.

Automatically called by the parser and ideally should be called manually after each tree modification.